Plumbing retry client into auth
