+++ src/auth/Cargo.toml
@@ -29,6 +29,7 @@ default                 = ["default-crypto-provider"]
 default-crypto-provider = ["rustls/aws_lc_rs"]
 
 [dependencies]
+retry          = { version = "0.1", path = "../retry", package = "google-cloud-retry", features = ["unstable-sdk-client"] }
 async-trait    = "0.1"
 http           = "1"
 reqwest        = { version = "0.12", features = ["json"] }
